Remove unnecessary type conversions

+++ pkg/controller/podautoscaler/metrics/rest_metrics_client.go
@@ -81,7 +81,7 @@ func (c *resourceMetricsClient) GetResourceMetric(resource v1.ResourceName, name
 		}
 
 		if !missing {
-			res[m.Name] = int64(podSum)
+			res[m.Name] = podSum
 		}
 	}
